chore: Renamed service/session-manager/session-manager.service.mts to service/session-manager/session.registry.mts.
