+++ packages/architectura/src/service/session-manager/_index.mts
@@ -1,1 +1,1 @@
-export * from './session-manager.service.mjs';
+export * from './session.registry.mjs';
