+++ packages/architectura/src/service/session-manager/session.registry.mts
